requirements.txt
# Core diffusion libraries
diffusers>=0.21.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.20.0

# Image processing and visualization
Pillow>=9.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.20.0

# Evaluation metrics - using compatible versions
open_clip_torch>=2.20.0
lpips>=0.1.4
pytorch-fid>=0.3.0

# Utilities
tqdm>=4.65.0
argparse
wandb>=0.15.0
tensorboard>=2.13.0

# Optional: for better performance
xformers>=0.0.20 